Move json tests to own file
Goal is to make replicache.test.ts smaller over time
