+++ src/replicache.test.ts
@@ -1,13 +1,10 @@
-import {
-  ReplicacheTest,
-  httpStatusUnauthorized,
-  MutatorDefs,
-} from './replicache.js';
+import {ReplicacheTest, httpStatusUnauthorized} from './replicache.js';
+import type {MutatorDefs} from './replicache.js';
 import type {ReplicacheOptions} from './replicache.js';
 import {Replicache, TransactionClosedError} from './mod.js';
 
 import type {ReadTransaction, WriteTransaction} from './mod.js';
-import {deepEqual, JSONValue} from './json.js';
+import type {JSONValue} from './json.js';
 
 import {assert, expect} from '@esm-bundle/chai';
 import * as sinon from 'sinon';
@@ -1886,57 +1883,6 @@ testWithBothStores('logLevel', async () => {
 
   // Restoring since we are not yet scoped to a Replicache db instance.
   rep = await replicacheForTesting('log-level', {logLevel: 'info'});
-});
-
-test('JSON deep equal', () => {
-  const t = (
-    a: JSONValue | undefined,
-    b: JSONValue | undefined,
-    expected = true,
-  ) => {
-    const res = deepEqual(a, b);
-    if (res !== expected) {
-      fail(
-        JSON.stringify(a) + (expected ? ' === ' : ' !== ') + JSON.stringify(b),
-      );
-    }
-  };
-
-  const oneLevelOfData = [
-    0,
-    1,
-    2,
-    3,
-    456789,
-    true,
-    false,
-    null,
-    '',
-    'a',
-    'b',
-    'cdefefsfsafasdadsaas',
-    [],
-    {},
-    {x: 4, y: 5, z: 6},
-    [7, 8, 9],
-  ] as const;
-
-  const testData = [
-    ...oneLevelOfData,
-    [...oneLevelOfData],
-    Object.fromEntries(oneLevelOfData.map(v => [JSON.stringify(v), v])),
-  ];
-
-  for (let i = 0; i < testData.length; i++) {
-    for (let j = 0; j < testData.length; j++) {
-      const a = testData[i];
-      // "clone" to ensure we do not end up with a and b being the same object.
-      const b = JSON.parse(JSON.stringify(testData[j]));
-      t(a, b, i === j);
-    }
-  }
-
-  t({a: 1, b: 2}, {b: 2, a: 1});
 });
 
 // Only used for type checking
